OneDrive/Escritorio/punto_mas/nuevo_prueba/src/home/api/views.py
from rest_framework import permissions, throttling, viewsets
from home import models


class SampleViewSet(viewsets.ModelViewSet):
    serializer_class = models.SampleSerializer
    queryset = models.Sample.objects.all()
    throttle_classes = [throttling.AnonRateThrottle]
    permission_classes = [
        permissions.IsAuthenticatedOrReadOnly,
    ]
    http_method_names = [
        "get",
    ]
