OneDrive/Escritorio/punto_mas/nuevo_prueba/pyproject.toml
[project]
description = "punto_mas"
name = "punto_mas"
readme = "README.md"
version = "0.0.1"

authors = [{name = "BigPear", email = "hola@bigpear.com.ar"}]
dependencies = [
  "Django==5.0.6",
  "django-cogwheels",
  "django-debug-toolbar",
  "django-environ",
  "django-extensions",
  "django-filter",
  "django-taggit",
  "django-vite",
  "djangorestframework",
  "django-cors-headers",
  "gunicorn",
]

[project.optional-dependencies]
dev = [
  "ruff",
  "cookiecutter",
  "coverage",
  "flake8",
  "flake8-pyproject",
  "ipdb",
  "pre-commit",
  "pytest",
  "pytest-django",
  "django-stubs",
  "wagtail-factories",
  "bump2version",
]
docs = [
  "mkdocs",
  "mkdocstrings",
  "mkdocstrings-python",
  "mkdocs-material",
  "mkdocs-material-extensions",
  "mkdocs-autolinks-plugin",
  "mkdocs-pdf-export-plugin",
  "mkdocs-print-site-plugin",
  "html5lib",
  "requests",
  "htmlark",
]

[project.urls]
documentation = "https://gitlab.com/bigpear.adm/desafio_future"
homepage = "https://gitlab.com/bigpear.adm/desafio_future"
repository = "https://gitlab.com/bigpear.adm/desafio_future"

[project.scripts]
punto_mas = "punto_mas.manage:do_manage"

[build-system]
build-backend = "setuptools.build_meta"
requires = ["setuptools"]

[tool.setuptools.packages.find]
where = ["src"]

[tool.isort]
known_first_party = [

  "home",
  "punto_mas",

]

[tool.flake8]
exclude = ['**/migrations/**']
extend-ignore = ["E203"]
ignore = ["E501", "F403", "E402", "E203", "W503", "Q000"]
max-complexity = 12
max-line-length = 120
statistics = true

[tool.ruff]
exclude = ["test/data"]
line-length = 120

[tool.ruff.lint]
# max-args = 20
# good-names = ["i", "j", "k", "ex", "_", "e", "db", "df", "ms", "ts"]
# ignore = ["E114", "C0115", "C0116","E0611", "R0902", "R0903", "R0913"]
ignore = ["D100", "D101", "D102", "D103"]

[tool.ruff.lint.flake8-unused-arguments]
ignore-variadic-names = true

[tool.mypy]
disable_error_code = [
  "attr-defined",
]
exclude = [
  "tests/data",
]
ignore_missing_imports = true
plugins = [
  "mypy_django_plugin.main",
  "mypy_drf_plugin.main",
]
python_version = "3.12"

[tool.django-stubs]
django_settings_module = "punto_mas.settings"

[tool.pytest.ini_options]
addopts = [
  "-p no:cacheprovider",
  "-ra",
  "--strict-markers",
  "--verbose",
  "--cov-config=.coveragerc",
  "--cov=src", #"--cov-report=term",
  "--cov-report=html:docs/coverage",
  "--cov-report=xml:docs/coverage/coverage.xml",
  "--junitxml=docs/test/junit/test-results.xml",
  "--verbose",
  "--cache-clear",
  "-m not slow",
]
markers = [
  "slow: marks tests as slow (deselect with '-m \"not slow\"')",
]
python_files = ["test_*.py"]
testpaths = [
  "tests",
]

[tool.coverage.run]
omit = []

[tool.coverage.report]
exclude_lines = [
  "pragma: no cover",
  "@abstractmethod",
]

[tool.pytest]
DJANGO_SETTINGS_MODULE = "punto_mas.settings"
python_files = "tests.py test_*.py *_tests.py"
